+++ source/visualization/OpenGL/src/G4OpenGLStoredSceneHandler.cc
@@ -1,12 +1,28 @@
-// This code implementation is the intellectual property of
-// the GEANT4 collaboration.
 //
-// By copying, distributing or modifying the Program (or any work
-// based on the Program) you indicate your acceptance of this statement,
-// and all its terms.
+// ********************************************************************
+// * DISCLAIMER                                                       *
+// *                                                                  *
+// * The following disclaimer summarizes all the specific disclaimers *
+// * of contributors to this software. The specific disclaimers,which *
+// * govern, are listed with their locations in:                      *
+// *   http://cern.ch/geant4/license                                  *
+// *                                                                  *
+// * Neither the authors of this software system, nor their employing *
+// * institutes,nor the agencies providing financial support for this *
+// * work  make  any representation or  warranty, express or implied, *
+// * regarding  this  software system or assume any liability for its *
+// * use.                                                             *
+// *                                                                  *
+// * This  code  implementation is the  intellectual property  of the *
+// * GEANT4 collaboration.                                            *
+// * By copying,  distributing  or modifying the Program (or any work *
+// * based  on  the Program)  you indicate  your  acceptance of  this *
+// * statement, and all its terms.                                    *
+// ********************************************************************
 //
-// $Id: G4OpenGLStoredSceneHandler.cc,v 1.9 2001/02/23 15:43:11 johna Exp $
-// GEANT4 tag $Name: geant4-03-01 $
+//
+// $Id: G4OpenGLStoredSceneHandler.cc,v 1.10.2.1 2001/06/28 19:15:43 gunter Exp $
+// GEANT4 tag $Name:  $
 //
 // 
 // Andrew Walkden  10th February 1997
@@ -111,7 +127,7 @@ void G4OpenGLStoredSceneHandler::ClearStore () {
 
   G4VSceneHandler::ClearStore ();  // Sets need kernel visit, etc.
 
-  int i;
+  size_t i;
 
   // Delete OpenGL display lists.
   for (i = 0; i < fPODLList.size (); i++) {
@@ -158,7 +174,7 @@ void G4OpenGLStoredSceneHandler::EndModeling () {
 //  }
 
   glNewList (fTopPODL, GL_COMPILE); {
-    for (int i = 0; i < fPODLList.size (); i++) {
+    for (size_t i = 0; i < fPODLList.size (); i++) {
       glPushMatrix();
       G4OpenGLTransform3D oglt (fPODLTransformList [i]);
       glMultMatrixd (oglt.GetGLMatrix ());
