+++ source/visualization/RayTracer/src/G4RayTrajectory.cc
@@ -1,12 +1,28 @@
-// This code implementation is the intellectual property of
-// the GEANT4 collaboration.
 //
-// By copying, distributing or modifying the Program (or any work
-// based on the Program) you indicate your acceptance of this statement,
-// and all its terms.
+// ********************************************************************
+// * DISCLAIMER                                                       *
+// *                                                                  *
+// * The following disclaimer summarizes all the specific disclaimers *
+// * of contributors to this software. The specific disclaimers,which *
+// * govern, are listed with their locations in:                      *
+// *   http://cern.ch/geant4/license                                  *
+// *                                                                  *
+// * Neither the authors of this software system, nor their employing *
+// * institutes,nor the agencies providing financial support for this *
+// * work  make  any representation or  warranty, express or implied, *
+// * regarding  this  software system or assume any liability for its *
+// * use.                                                             *
+// *                                                                  *
+// * This  code  implementation is the  intellectual property  of the *
+// * GEANT4 collaboration.                                            *
+// * By copying,  distributing  or modifying the Program (or any work *
+// * based  on  the Program)  you indicate  your  acceptance of  this *
+// * statement, and all its terms.                                    *
+// ********************************************************************
 //
-// $Id: G4RayTrajectory.cc,v 1.6 2001/02/23 15:43:14 johna Exp $
-// GEANT4 tag $Name: geant4-03-01 $
+//
+// $Id: G4RayTrajectory.cc,v 1.7.2.1 2001/06/28 19:15:54 gunter Exp $
+// GEANT4 tag $Name:  $
 //
 //
 //
@@ -37,7 +53,7 @@ G4RayTrajectory :: G4RayTrajectory()
 G4RayTrajectory :: G4RayTrajectory(G4RayTrajectory & right)
 {
   positionRecord = new G4std::vector<G4RayTrajectoryPoint*>;
-  for(int i=0;i<right.positionRecord->size();i++)
+  for(size_t i=0;i<right.positionRecord->size();i++)
   {
     G4RayTrajectoryPoint* rightPoint = (G4RayTrajectoryPoint*)
 				((*(right.positionRecord))[i]);
@@ -48,7 +64,7 @@ G4RayTrajectory :: G4RayTrajectory(G4RayTrajectory & right)
 G4RayTrajectory :: ~G4RayTrajectory()
 {
   //positionRecord->clearAndDestroy();
-  for(int i=0;i<positionRecord->size();i++)
+  for(size_t i=0;i<positionRecord->size();i++)
   { delete (*positionRecord)[i]; }
   positionRecord->clear();
   delete positionRecord;
